Refactor: DTO 리패키징
HTTP 요청/응답, 쿼리 응답 기준으로 추가적인 하위 패키징
