+++ src/main/java/com/wanted/jaringoby/domains/category/dtos/http/GetCategoryListResponseDto.java
@@ -1,4 +1,4 @@
-package com.wanted.jaringoby.domains.category.dtos;
+package com.wanted.jaringoby.domains.category.dtos.http;
 
 import java.util.List;
 import lombok.Builder;
