+++ src/main/java/com/wanted/jaringoby/domains/customer/applications/CreateCustomerService.java
@@ -1,8 +1,8 @@
 package com.wanted.jaringoby.domains.customer.applications;
 
 import com.wanted.jaringoby.common.utils.UlidGenerator;
-import com.wanted.jaringoby.domains.customer.dtos.CreateCustomerRequestDto;
-import com.wanted.jaringoby.domains.customer.dtos.CreateCustomerResponseDto;
+import com.wanted.jaringoby.domains.customer.dtos.http.CreateCustomerRequestDto;
+import com.wanted.jaringoby.domains.customer.dtos.http.CreateCustomerResponseDto;
 import com.wanted.jaringoby.domains.customer.entities.Customer;
 import com.wanted.jaringoby.domains.customer.exceptions.CustomerReconfirmPasswordMismatchedException;
 import com.wanted.jaringoby.domains.customer.exceptions.CustomerUsernameDuplicatedException;
