+++ src/main/java/com/wanted/jaringoby/domains/customer/controllers/CustomerController.java
@@ -4,8 +4,8 @@ import com.wanted.jaringoby.common.response.Response;
 import com.wanted.jaringoby.common.validations.BindingResultChecker;
 import com.wanted.jaringoby.common.validations.ValidationSequence;
 import com.wanted.jaringoby.domains.customer.applications.CreateCustomerService;
-import com.wanted.jaringoby.domains.customer.dtos.CreateCustomerRequestDto;
-import com.wanted.jaringoby.domains.customer.dtos.CreateCustomerResponseDto;
+import com.wanted.jaringoby.domains.customer.dtos.http.CreateCustomerRequestDto;
+import com.wanted.jaringoby.domains.customer.dtos.http.CreateCustomerResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
