+++ src/main/java/com/wanted/jaringoby/domains/ledger/applications/BudgetRecommendationService.java
@@ -3,8 +3,8 @@ package com.wanted.jaringoby.domains.ledger.applications;
 import com.wanted.jaringoby.common.models.Money;
 import com.wanted.jaringoby.common.models.Percentage;
 import com.wanted.jaringoby.domains.category.models.Category;
-import com.wanted.jaringoby.domains.ledger.dtos.GetBudgetRecommendationQueryParamsDto;
-import com.wanted.jaringoby.domains.ledger.dtos.GetBudgetRecommendationResponseDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.GetBudgetRecommendationQueryParamsDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.GetBudgetRecommendationResponseDto;
 import com.wanted.jaringoby.domains.ledger.dtos.query.LedgerIdAndBudgetsQueryResultDto;
 import com.wanted.jaringoby.domains.ledger.entities.budget.Budget;
 import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
