+++ src/main/java/com/wanted/jaringoby/domains/ledger/applications/ModifyLedgerBudgetsService.java
@@ -6,8 +6,8 @@ import com.wanted.jaringoby.domains.category.exceptions.CategoryDuplicatedExcept
 import com.wanted.jaringoby.domains.category.exceptions.CategoryNotFoundException;
 import com.wanted.jaringoby.domains.category.models.Category;
 import com.wanted.jaringoby.domains.customer.entities.CustomerId;
-import com.wanted.jaringoby.domains.ledger.dtos.BudgetRequestDto;
-import com.wanted.jaringoby.domains.ledger.dtos.ModifyLedgerBudgetsRequestDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.BudgetRequestDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.ModifyLedgerBudgetsRequestDto;
 import com.wanted.jaringoby.domains.ledger.entities.budget.Budget;
 import com.wanted.jaringoby.domains.ledger.entities.ledger.Ledger;
 import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
