+++ src/main/java/com/wanted/jaringoby/domains/ledger/dtos/http/CreateLedgerResponseDto.java
@@ -1,4 +1,4 @@
-package com.wanted.jaringoby.domains.ledger.dtos;
+package com.wanted.jaringoby.domains.ledger.dtos.http;
 
 import lombok.Builder;
