+++ src/main/java/com/wanted/jaringoby/domains/ledger/entities/budget/Budget.java
@@ -4,7 +4,7 @@ import com.wanted.jaringoby.common.converters.MoneyConverter;
 import com.wanted.jaringoby.common.models.Money;
 import com.wanted.jaringoby.common.models.Percentage;
 import com.wanted.jaringoby.domains.category.models.Category;
-import com.wanted.jaringoby.domains.ledger.dtos.GetBudgetResponseDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.GetBudgetResponseDto;
 import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
 import jakarta.persistence.AttributeOverride;
 import jakarta.persistence.Column;
