+++ src/main/java/com/wanted/jaringoby/session/applications/LoginService.java
@@ -5,8 +5,8 @@ import com.wanted.jaringoby.common.utils.UlidGenerator;
 import com.wanted.jaringoby.domains.customer.entities.Customer;
 import com.wanted.jaringoby.domains.customer.exceptions.CustomerNotFoundException;
 import com.wanted.jaringoby.domains.customer.repositories.CustomerRepository;
-import com.wanted.jaringoby.session.dtos.LoginRequestDto;
-import com.wanted.jaringoby.session.dtos.LoginResponseDto;
+import com.wanted.jaringoby.session.dtos.http.LoginRequestDto;
+import com.wanted.jaringoby.session.dtos.http.LoginResponseDto;
 import com.wanted.jaringoby.session.entities.CustomerRefreshToken;
 import com.wanted.jaringoby.session.exceptions.CustomerPasswordMismatchedException;
 import com.wanted.jaringoby.session.repositories.CustomerRefreshTokenRepository;
