+++ src/main/java/com/wanted/jaringoby/session/applications/ReissueAccessTokenService.java
@@ -2,7 +2,7 @@ package com.wanted.jaringoby.session.applications;
 
 import com.wanted.jaringoby.common.utils.JwtUtil;
 import com.wanted.jaringoby.domains.customer.entities.CustomerId;
-import com.wanted.jaringoby.session.dtos.ReissueAccessTokenResponseDto;
+import com.wanted.jaringoby.session.dtos.http.ReissueAccessTokenResponseDto;
 import com.wanted.jaringoby.session.exceptions.CustomerRefreshTokenNotFoundException;
 import com.wanted.jaringoby.session.repositories.CustomerRefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
