+++ src/main/java/com/wanted/jaringoby/session/controllers/SessionController.java
@@ -5,8 +5,8 @@ import com.wanted.jaringoby.common.validations.BindingResultChecker;
 import com.wanted.jaringoby.common.validations.ValidationSequence;
 import com.wanted.jaringoby.session.applications.LoginService;
 import com.wanted.jaringoby.session.applications.LogoutService;
-import com.wanted.jaringoby.session.dtos.LoginRequestDto;
-import com.wanted.jaringoby.session.dtos.LoginResponseDto;
+import com.wanted.jaringoby.session.dtos.http.LoginRequestDto;
+import com.wanted.jaringoby.session.dtos.http.LoginResponseDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
