+++ src/main/java/com/wanted/jaringoby/session/dtos/http/LoginResponseDto.java
@@ -1,4 +1,4 @@
-package com.wanted.jaringoby.session.dtos;
+package com.wanted.jaringoby.session.dtos.http;
 
 import lombok.Builder;
