+++ src/test/java/com/wanted/jaringoby/domains/customer/applications/CreateCustomerServiceTest.java
@@ -9,8 +9,8 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 
 import com.wanted.jaringoby.common.utils.UlidGenerator;
-import com.wanted.jaringoby.domains.customer.dtos.CreateCustomerRequestDto;
-import com.wanted.jaringoby.domains.customer.dtos.CreateCustomerResponseDto;
+import com.wanted.jaringoby.domains.customer.dtos.http.CreateCustomerRequestDto;
+import com.wanted.jaringoby.domains.customer.dtos.http.CreateCustomerResponseDto;
 import com.wanted.jaringoby.domains.customer.entities.Customer;
 import com.wanted.jaringoby.domains.customer.exceptions.CustomerReconfirmPasswordMismatchedException;
 import com.wanted.jaringoby.domains.customer.exceptions.CustomerUsernameDuplicatedException;
