+++ src/test/java/com/wanted/jaringoby/domains/customer/controllers/CustomerControllerTest.java
@@ -12,8 +12,8 @@ import com.wanted.jaringoby.common.config.security.SecurityConfig;
 import com.wanted.jaringoby.common.config.validation.ValidationConfig;
 import com.wanted.jaringoby.common.validations.BindingResultChecker;
 import com.wanted.jaringoby.domains.customer.applications.CreateCustomerService;
-import com.wanted.jaringoby.domains.customer.dtos.CreateCustomerRequestDto;
-import com.wanted.jaringoby.domains.customer.dtos.CreateCustomerResponseDto;
+import com.wanted.jaringoby.domains.customer.dtos.http.CreateCustomerRequestDto;
+import com.wanted.jaringoby.domains.customer.dtos.http.CreateCustomerResponseDto;
 import com.wanted.jaringoby.domains.customer.repositories.CustomerRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
