+++ src/test/java/com/wanted/jaringoby/domains/ledger/applications/ModifyLedgerBudgetsServiceTest.java
@@ -16,8 +16,8 @@ import com.wanted.jaringoby.common.utils.UlidGenerator;
 import com.wanted.jaringoby.domains.category.exceptions.CategoryDuplicatedException;
 import com.wanted.jaringoby.domains.category.exceptions.CategoryNotFoundException;
 import com.wanted.jaringoby.domains.category.models.Category;
-import com.wanted.jaringoby.domains.ledger.dtos.BudgetRequestDto;
-import com.wanted.jaringoby.domains.ledger.dtos.ModifyLedgerBudgetsRequestDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.BudgetRequestDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.ModifyLedgerBudgetsRequestDto;
 import com.wanted.jaringoby.domains.ledger.entities.budget.Budget;
 import com.wanted.jaringoby.domains.ledger.entities.ledger.Ledger;
 import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
