+++ src/test/java/com/wanted/jaringoby/domains/ledger/applications/ModifyLedgerPeriodServiceTest.java
@@ -9,7 +9,7 @@ import static org.mockito.Mockito.verify;
 
 import com.wanted.jaringoby.common.constants.Date;
 import com.wanted.jaringoby.domains.customer.entities.CustomerId;
-import com.wanted.jaringoby.domains.ledger.dtos.ModifyLedgerPeriodRequestDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.ModifyLedgerPeriodRequestDto;
 import com.wanted.jaringoby.domains.ledger.entities.ledger.Ledger;
 import com.wanted.jaringoby.domains.ledger.entities.ledger.LedgerId;
 import com.wanted.jaringoby.domains.ledger.exceptions.LedgerNotFoundException;
