+++ src/test/java/com/wanted/jaringoby/domains/ledger/controllers/LedgerControllerTest.java
@@ -24,15 +24,15 @@ import com.wanted.jaringoby.domains.ledger.applications.CreateLedgerService;
 import com.wanted.jaringoby.domains.ledger.applications.GetOngoingLedgerService;
 import com.wanted.jaringoby.domains.ledger.applications.ModifyLedgerBudgetsService;
 import com.wanted.jaringoby.domains.ledger.applications.ModifyLedgerPeriodService;
-import com.wanted.jaringoby.domains.ledger.dtos.BudgetRecommendationDto;
-import com.wanted.jaringoby.domains.ledger.dtos.CreateLedgerRequestDto;
-import com.wanted.jaringoby.domains.ledger.dtos.CreateLedgerResponseDto;
-import com.wanted.jaringoby.domains.ledger.dtos.GetBudgetRecommendationQueryParamsDto;
-import com.wanted.jaringoby.domains.ledger.dtos.GetBudgetRecommendationResponseDto;
-import com.wanted.jaringoby.domains.ledger.dtos.GetBudgetResponseDto;
-import com.wanted.jaringoby.domains.ledger.dtos.GetLedgerDetailResponseDto;
-import com.wanted.jaringoby.domains.ledger.dtos.ModifyLedgerBudgetsRequestDto;
-import com.wanted.jaringoby.domains.ledger.dtos.ModifyLedgerPeriodRequestDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.BudgetRecommendationDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.CreateLedgerRequestDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.CreateLedgerResponseDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.GetBudgetRecommendationQueryParamsDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.GetBudgetRecommendationResponseDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.GetBudgetResponseDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.GetLedgerDetailResponseDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.ModifyLedgerBudgetsRequestDto;
+import com.wanted.jaringoby.domains.ledger.dtos.http.ModifyLedgerPeriodRequestDto;
 import java.time.LocalDate;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
