+++ src/test/java/com/wanted/jaringoby/session/controllers/SessionControllerTest.java
@@ -17,8 +17,8 @@ import com.wanted.jaringoby.domains.customer.entities.CustomerId;
 import com.wanted.jaringoby.domains.customer.repositories.CustomerRepository;
 import com.wanted.jaringoby.session.applications.LoginService;
 import com.wanted.jaringoby.session.applications.LogoutService;
-import com.wanted.jaringoby.session.dtos.LoginRequestDto;
-import com.wanted.jaringoby.session.dtos.LoginResponseDto;
+import com.wanted.jaringoby.session.dtos.http.LoginRequestDto;
+import com.wanted.jaringoby.session.dtos.http.LoginResponseDto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
